# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Restaurant Finder API
  version: 1.0.0
  description: API for searching restaurants, menu items, and popular dishes.
servers:
  - url: http://localhost:5000
paths:
  /restaurants:
    get:
      summary: Search all restaurants or optionally filter by cuisine
      parameters:
        - name: cuisine
          in: query
          description: Cuisine type to filter restaurants
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of restaurants
          content:
            application/json:
              schema:
                type: object
                properties:
                  restaurants:
                    type: array
                    items:
                      $ref: '#/components/schemas/Restaurant'
  /restaurant/{id}:
    get:
      summary: Get restaurant details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Restaurant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
  /restaurant:
    post:
      summary: Add a new restaurant record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Restaurant'
      responses:
        '201':
          description: Restaurant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'

  /restaurant/{id}/menu:
    post:
      summary: Add a new menu item for a restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

  /menu-item:
    post:
      summary: Add a new menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuItemInput'
      responses:
        '201':
          description: Menu item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItem'
  /restaurant/{id}/popular:
    post:
      summary: Add a new popular dish for a restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

  /popular-dish:
    post:
      summary: Add a new popular dish
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PopularDishInput'
      responses:
        '201':
          description: Popular dish created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PopularDish'
components:
  schemas:
    Restaurant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        website:
          type: string
        cuisine:
          type: string
        latitude:
          type: number
        longitude:
          type: number
    RestaurantInput:
      type: object
      required:
        - name
        - address
        - phone
        - cuisine
        - website
        - latitude
        - longitude
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        website:
          type: string
        cuisine:
          type: string
        latitude:
          type: number
        longitude:
          type: number
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
    MenuItemInput:
      type: object
      required:
        - name
        - description
        - price
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
    PopularDish:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    PopularDishInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string